refactor: import from text files rather than from a CSV
