+++ README.md
@@ -13,7 +13,7 @@ Options:
 - `--get-by-tag <tag>`: screeny will retrieve 10 articles from Ophan by the given tag and use these URLs in the requests to urlBox
 - `--get-for-all-tags`: screeny will retrieve 10 articles for each tag from Ophan and use these URLs in the requests to urlBox
 - `--import-from-google-sheets <spreadsheetId> <sheetId>`: this option requires the user to provide both the spreadsheetId and sheetId (if not the default). Screeny will authenticate and then read all the URLs from the provided Google sheet (assuming URLs are in range A2:A)
-- `--import from csv <relative path>` screeny will read all entries from a CSV, formatted as the one shown in docs/example.csv
+- `--import-from-file <relative path>` screeny will read all entries from  plane text file, formatted as the one shown in docs/example.txt
 - `--compare-to-dcr`: given this option screeny will formulate a second URL from those given (either retrieved from Ophan or Google sheets). By appending `?dcr` to the URL the article will be rendered via DCR. This option doubles the number of requests to urlBox and the output will be side-by-side images comparing articles rendered via DCR compared to Frontend.
 
 Example request to import URLs from the example CSV and generate side-by-side images comparing the page rendered via frontend to DCR:
@@ -21,5 +21,5 @@ Example request to import URLs from the example CSV and generate side-by-side im
 ```
 nvm use
 yarn
-node --max-old-space-size=25000 index.js --import-from-csv "./docs/example.csv" --compare-to-dcr
+node --max-old-space-size=25000 index.js --import-from-file "./docs/example.txt" --compare-to-dcr
 ```
