+++ docs/example.txt
@@ -1,0 +1,5 @@
+https://www.theguardian.com/film/2023/mar/13/oscars-2023-from-brendan-fraser-to-ke-huy-quan-hollywood-loves-a-comeback-story
+https://www.theguardian.com/education/2023/mar/13/sheffield-university-criticised-over-handling-of-protest-investigation
+https://www.theguardian.com/food/2023/mar/13/torta-del-paradiso-paradise-cake-recipe-rachel-roddy
+https://www.theguardian.com/food/2023/mar/12/jay-rayner-restaurant-review-isibani-london-bathe-me-in-the-pepper-soup
+invalid URL example
