+++ package.json
@@ -10,7 +10,6 @@
   "dependencies": {
     "axios": "^0.21.3",
     "commander": "^8.1.0",
-    "csv-parse": "^5.3.6",
     "dotenv": "^10.0.0",
     "is-plain-obj": "^4.1.0",
     "jimp": "^0.22.7",
